Cleaned up Work transaction and record comments.
