+++ src/tests/mongo/test_mongo_persists.py
@@ -6,7 +6,7 @@ from commonworks.agreement import AgreementRecord, AgreementInterestedParty
 from commonworks.interested_party import Publisher
 from commonworks.value_entity import ValueEntity
 from commonworks.work import AlternateTitle, AuthoredWork, \
-    PerformingArtist, WorkOrigin, Work, RecordingDetail
+    PerformingArtist, WorkOrigin, WorkRecord, RecordingDetail
 from tests.mongo.mongo_test_conf import host, port, db_name, MongoGenericRepository
 
 
@@ -74,7 +74,7 @@ class TestAuthoredWork(unittest.TestCase):
     """
 
     def setUp(self):
-        self.entity = AuthoredWork(1, 'title', 'ES', 'Broadway show',
+        self.entity = AuthoredWork(None, 1, 'title', 'ES', 'Broadway show',
                                    'name1', 1, 'ip_1',
                                    'name2', 2, 'ip_2', 'surname1', 'surname2',
                                    3)
@@ -187,17 +187,17 @@ class TestWork(unittest.TestCase):
     """
 
     def setUp(self):
-        self.entity = Work(1, 'The Title', 'original', 'category',
-                           language_code='ES', printed_edition_publication_date=datetime.date(2015, 1, 11),
-                           copyright_number=2, copyright_date=datetime.date(2015, 1, 12),
-                           text_music_relationship='text_only',
-                           music_arrangement='none', lyric_adaptation='none', excerpt_type='movement',
-                           composite_type='composite', composite_component_count=3,
-                           iswc=4, cwr_work_type='jazz',
-                           duration=60, catalogue_number=5, opus_number='28#3',
-                           contact_id='name_id', contact_name='Person',
-                           recorded_indicator=True, priority_flag=True, exceptional_clause=True,
-                           grand_rights_indicator=True)
+        self.entity = WorkRecord(None, 1, 'The Title', 'original', 'category',
+                                 language_code='ES', printed_edition_publication_date=datetime.date(2015, 1, 11),
+                                 copyright_number=2, copyright_date=datetime.date(2015, 1, 12),
+                                 text_music_relationship='text_only',
+                                 music_arrangement='none', lyric_adaptation='none', excerpt_type='movement',
+                                 composite_type='composite', composite_component_count=3,
+                                 iswc=4, cwr_work_type='jazz',
+                                 duration=60, catalogue_number=5, opus_number='28#3',
+                                 contact_id='name_id', contact_name='Person',
+                                 recorded_indicator=True, priority_flag=True, exceptional_clause=True,
+                                 grand_rights_indicator=True)
         self.repo = MongoGenericRepository(host, port, db_name, 'works')
 
     def tearDown(self):
